+++ gui_window.cpp
@@ -30,5 +30,5 @@ float GUIElementWindow::GetAlpha() const {
 
 void GUIElementWindow::RenderElement(Texture *texturer, Render *renderer, const byte &active, const byte &hover) const {
 	texturer->Bind(texturer->GetWhite());
-	renderer->DrawRect(pixelPos.GetX(), renderer->GetScreenHeight() - pixelPos.GetY(), pixelSize.GetX(), pixelSize.GetY(), 0.0f, 0.0f, 0.0f, alpha);
+	renderer->DrawRect((float)pixelPos.GetX(), (float)(renderer->GetScreenHeight() - pixelPos.GetY()), (float)pixelSize.GetX(), (float)pixelSize.GetY(), 0.0f, 0.0f, 0.0f, alpha);
 }
